+++ init/40_firstrun.sh
@@ -6,10 +6,6 @@
 # Github URL for opencv zip file download.
 # Current default is to pull the version 4.3.0 release.
 #
-OPENCV_VER=4.3.0
-OPENCV_URL=https://github.com/opencv/opencv/archive/$OPENCV_VER.zip
-OPENCV_CONTRIB_URL=https://github.com/opencv/opencv_contrib/archive/$OPENCV_VER.zip
-
 # Search for config files, if they don't exist, create the default ones
 if [ ! -d /config/conf ]; then
 	echo "Creating conf folder"
@@ -361,38 +357,16 @@ if [ "$INSTALL_HOOK" == "1" ]; then
 		# pip3 will take care of installing dependent packages
 		pip3 install future
 		pip3 install /root/zmeventnotification
-		#
+
+
 		# Compile opencv
-		#
-		echo "Compiling opencv..."
-		cd ~
-		wget -q -O opencv.zip $OPENCV_URL
-		wget -q -O opencv_contrib.zip $OPENCV_CONTRIB_URL
-		unzip opencv.zip
-		unzip opencv_contrib.zip
-		mv $(ls -d opencv-*) opencv
-		mv opencv_contrib-$OPENCV_VER opencv_contrib
-		rm *.zip
+		if [ ! -f /root/openccv_compile.sh ]; then
+			echo "Compiling opencv - this will take a while..."
+			chmod +x /root/opencv_compile.sh
+			/root/opencv_compile.sh > /dev/null
+		fi
 
-		cd ~/opencv
-		mkdir build
-		cd build
-		cmake -D CMAKE_BUILD_TYPE=RELEASE \
-			-D CMAKE_INSTALL_PREFIX=/usr/local \
-			-D INSTALL_PYTHON_EXAMPLES=OFF \
-			-D INSTALL_C_EXAMPLES=OFF \
-			-D OPENCV_ENABLE_NONFREE=ON \
-			-D OPENCV_EXTRA_MODULES_PATH=~/opencv_contrib/modules \
-			-D HAVE_opencv_python3=ON \
-			-D PYTHON_EXECUTABLE=/usr/bin/python3 \
-			-D PYTHON2_EXECUTABLE=/usr/bin/python2 \
-			-D BUILD_EXAMPLES=OFF .. >/config/opencv/cmake.log
-
-		echo "Installing opencv..."
-		make -j$(nproc)
-		make install
 		cd ~
-		rm -r opencv*
 	    rm -rf /root/zmeventnotification/zmes_hook_helpers
 	fi